Feature room

+++ app/src/main/java/ucne/edu/proyectofinalaplicada2/data/remote/dto/RentaDto.kt
@@ -6,5 +6,5 @@ data class RentaDto(
     val vehiculoId: Int?,
     val fechaRenta: String?,
     val fechaEntrega: String?,
-    val total: Int?
+    val total: Double?
 )
